requirements.txt
django==5.0.6
djangorestframework==3.15.1
django-environ==0.11.2
django-filter==24.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
